Add functionality to use a duplicity pre_command
The duplicity module supports a pre_command, to be
run before a duplicity backup. We want duplicity
to be able to use the private GPG key for
encrypting backups, which has a passphrase.
Duplicity uses env variable `PASSPHRASE` for the
private GPG key. We will use the pre_command to
export this env variable for duplicity.