Environment variable to allow any HMRC manual slug
The HMRC Manuals API has a whitelist of known valid manual slugs. Recent
changes to that app (https://github.com/alphagov/hmrc-manuals-api/pull/112)
prevent publishing an HMRC Manual with a slug that isn't on the list.

But using the `ALLOW_UNKNOWN_HMRC_MANUAL_SLUGS` environment variable, we can
allow the app to publish manuals with any slug at all.

This is only technically necessary for Preview* but we've set it for
Development, too, for consistency with the recently added `publish_topics`.

* HMRC Manuals API only uses this environment variable when it's in
  a production environment. To check this, it uses `Rails.env.production?`,
  which is also `true` in Preview.