Add a govuk-jenkins aptly repo (not mirror)
The Jenkins project publish an apt repo with packages for Jenkins,
however this only contains the latest version (or latest LTS version).
Because of this using an aptly mirror means that old versions get
removes as new ones are added.

Currently CI is running an older version of Jenkins than our deploy
boxes. This version is no longer available in either the Jenkins apt
repo, or our mirror of it. This means that it's not possible to
provision ci-master without manual intervention.

Old packages are available for download from the Jenkins site[1] though,
so this PR creates an aptly repo that we can manually insert the jenkins
versions we need into, and then publish. This will allow us much more
control over the versions we retain. Once this has been populated, both
CI, and the deploy machines can be pointed at it, and the existing
mirror of the jenkins repo can be removed.

[1] http://pkg.jenkins-ci.org/debian-stable/