Put CDN logs in the deploy group
This is needed so that govuk-cdn-logs-monitor can read the log files as
they're being written.

In order to do this, we need to move from the "legacy" configuration
format, to "RainerScript".  This is now the recommended way to
configuration actions, according to the rsyslog documentation.

A significant difference with this configuration method is that the
global file permissions, etc, aren't used, so we have to specify them
all explicitly here.  This is probably good for clarity, anyway.

The configuration directives used in the "omfile" action are documented
here: http://www.rsyslog.com/doc/v8-stable/configuration/modules/omfile.html

I've tested this in vagrant-govuk, and it creates the log files with the
desired group (and the other permissions as before).