Create passive check jobs
We need a way to alert Icinga for Jenkins job. This commit adds two
jobs, one for SUCCESS and one for FAILURE. In Jenkins, you can add a
parameterized build as a Post-build Action which will run on certain
conditions. Due to the number of different conditions that can be set, I
thought having two separate jobs that are triggered on different
conditions would be better.

For example, you could trigger the `Success_Passive_Check` job on the
`Stable or unstable but not failed` condition.

You could then trigger the `Failure_Passive_Check` job on the `Failed`
condition.

There are two build parameters that can be entered:
`NCSA_CHECK_DESCRIPTION`
`NCSA_OUTPUT`

The description is what it says on the tin, and the output could be
useful if you wanted to alert for several different failure conditions
i.e. you could have different alerts for an unstable build and a
completely failed build.