Add Icinga passive checks to jobs
Previously the required triggers were added to jobs, but because the
Icinga defined services didn't exist they were not able to actually
report on the passive check, throwing the following error:

```
/var/log/icinga/nagios.log:[1445257363] EXTERNAL COMMAND:
PROCESS_SERVICE_CHECK_RESULT;10.3.0.3;"Data sync to Staging";0;"Data
sync to Staging successful"
/var/log/icinga/nagios.log:[1445257363] Warning:  Passive check result
was received for service '"Data sync to Staging"' on host '10.3.0.3',
but the service could not be found!
/var/log/icinga/nagios.log:[1445257363] Error: External command failed
-> PROCESS_SERVICE_CHECK_RESULT;10.3.0.3;"Data sync to Staging";0;"Data
sync to Staging successful"
```

This commit looks to add the Icinga passive check per job that you want
to be alerted on, and also attempt to template in the service
description within the template.