Bugfix: fix Jenkins jobs passive checks
There was an issue with how the logic denoted `$title` in this manifest,
so I have updated with just using a string.

I tried to do something clever with substituting spaces for underscores
for `$check_name` and had problems, so just went with adding a string in
the end. This might be better for flexibility and can always be changed
per job anyway.