Run node spec tests in parallel
Run the node spec tests in parallel, using one process for each CPU on
the machine running the tests. By doing this, I hope to run these tests
as part of our normal CI builds, which was previously prohibitive due to
the time they took.

We take the list of all node classes and divide it so that each process
runs a subset of the classes.

The code is inspired by this blog post:
http://blog.pivotal.io/labs/labs/parallelize-your-rspec-suite

...which also inspired the `parallel_tests` gem which we already use in
this codebase.

I tried to run the tests in parellel using the `parallel_tests` gem or
the `parallel_split_test` gem but neither provide the primitives to run
the same test example, from the same file, in parallel and with
different fixtures (i.e. Hiera data).

I also looked to see if Rspec support some kind of dynamic generation of
test assertions that would allow me to use `parallel_test`, but could
not find any way to do that.

On my local machine, running the tests in parallel takes 3 minutes 56
seconds, compared to 25 minutes 4 seconds sequentially.

I have no reason to believe that running the tests in parallel causes
any problems due to cross-contamination between the tests; anecdotal
evidence during testing backs this up.

I'm using fetch when looking up `ENV['classes']` so that it fails
noisily if the environment variable is not set.