Reserve port for GOV.UK Component Guide
The component guide app [1] isn't part of our production stack, but it
would be useful to give it a fixed port, rather than having to run on
`3000` and risk conflicting with other rails apps.

I've had this with govuk-prototyping-rails, which is common to run along
side the component guide.

Reserve a port here, so that it can use a port that won't clash with
another GOV.UK app in the future

[1] https://github.com/alphagov/govuk-component-guide