Raise Spotlight app alert threshold
Story:
https://trello.com/c/6v3ErDVB/10-raise-the-threshold-for-apps-that-frequently-500

Spotlight has generated spurious HTTP 5XX errors for some time.

Having spoken to @timothymower, it seems that these errors occur when
building a HTML DOM tree, which is a fundamental part of the
application's functionality and cannot be fixed easily.

Until such time as the application can be made more resilient, increase
the alert threshold so that it does not alert Second Line when they are
unlikely to be able to take remedial action.