Set MongoDB ENV variable for asset_manager
`asset_manager` is being upgrade to Rails 4, which involved upgrading
to `mongoid` 4, which requires a change in how the config is structured.

See https://github.com/alphagov/asset-manager/pull/42

Rather than update the existing `mongoid.yml` on `alphagov-deployment`
this moves to a 12 Factor approach, using environment variables, to
be consume by the config on the main app repo.

See https://github.gds/gds/alphagov-deployment/pull/1099 for old
approach, and note

- this no longer specifies the port, i assume picks `27017` by default
- the `refresh_interval` option will need to be moved into the public
  config, but i don't see that being an issue.

Once this is merged, and the public `mongoid.yml` is updated to user
ENV vars, and set `refresh_interval` we should be able to remove
the `alphagov-deployment` file.

Based on https://github.gds/gds/puppet/pull/3465/files