Fix base64 encoding in Pingdom check script
The `str.encode('base64')` method will add a newline character every
76 characters which causes the HTTP authentication header to break.

We discovered this because we increased the length of the username
that we're using with Pingdom, which increased the length of the base64
encoding.

`base64.b64encode()` doesn't have this problem and also doesn't need
the trailing newline to be stripped.