Change api-mongo replicaset config
In 80026c5505b8ff02cb70f77676648dacad9456d3 this was moved to be
environment-specific, which is correct, but the node specs also
require this hieradata to be in place.

Moving the preview values back into the common configuration file
and then overriding them in staging and production will fix it.