Fix mongo cluster bootstrapping.
The new configure-node-priority script can only work if the mongo
clustering has already been configured by the congifure-replica-script
(it errors if this hasn't been done). There's nothing currently
enforcing this order, so this can break when provisioning a new cluster
(eg on the dev VM which is configured as a single-node cluster).

Adding a require forces these to scripts to run in the required order.