Remove notices to say files are managed by Puppet
All of our system files are managed by Puppet if we have
modified them, so it's not helpful for us to add notices saying as much,
especially as we have been inconsistent in doing this.

External modules that we are using may still add these notices, but
there's nothing we can do about that.

This will need careful deploying to ensure that MySQL does not restart
due to the change to `debian.cnf.erb`; I don't think it will.