Rename class/define spec files to match convention
We use a convention elsewhere in this code base where the name of the
spec files for Puppet classes or defined types are as follows:

class__subclass_spec.rb

Where `class__subclass` corresponds to `class::subclass`, so the two
colons are replaced by two underscores.

This commits makes the filenames consistent.