Always run node spec tests
Now that the spec tests to ensure that node manifests compile
successfully are run in parallel as of 8709aa5, include them in the
default Rake task so that they run as part of our CI tests and when
`bundle exec rake` is invoked on a developer's machine.

The node specs take between 2-5 minutes to run depending on the speed
your machine, but I think the time tradeoff is worthwhile to avoid
errors making it into master.