Create a Jenkins job to trigger the Data_Sync_Complete jobs
At the moment the Data_Sync_Complete jobs are only triggered if the data sync
job completes entirely successfully; if one part of the sync fails then the
cleanup tasks for other, successful, parts don't get run. If we notice that
this has happened then we usually run the Data_Sync_Complete job manually,
because that should be idempotent and we need to fix problems which can be
caused by the sync.

We want to ensure that the Data_Sync_Complete jobs are always triggered by the
sync ending so that Preview and Staging work again as soon as possible and
alerts related to the sync are cleared.

This commit adds a separate job which can trigger the Data_Sync_Complete job
in either Staging or Preview so that we can put it downstream of the sync
jobs. I can't see a way of keeping it as part of the sync job and still have
it run if an earlier part of the sync job has failed.

I've had to make this job a parameterized build because only
`trigger-parameterized-builds` allows `condition` to be set to ALWAYS;
`trigger` only allows SUCCESS, UNSTABLE or FAILURE. (I assume that
`trigger-parameterized-builds` can only be used with jobs which have
parameters.) Having a single job which can do this for either environment is a
bit less messy than having a separate one for each anyway.