Add a defined type for MONGODB_URI envvar
We're already creating the MONGODB_URI envvar for 5 apps and we're about to
add it for 2 more, so this seems like a good time to create a defined type for
it to avoid further duplication.

This is fairly similar to the existing database_uri defined type, but has a
couple of differences:

- it expects and validates for a non-empty array of hosts
- it doesn't use a username and password

Given those differences I didn't think it was worth the increase in complexity
to try to make database_url handle this case as well, so created a separate
defined type for it.