Remove api-mongo-4 from production replica members
This node does not currently exist and this is
causing issues when we try to configure node
priority. We use rs.config() to get priority of
an expected set of nodes. We don’t want to check
for existence of a node first and skip if not 
present as this could cause weird issues. Instead,
remove from hieradata for now so it is not in
expected replica set members. See:
https://github.gds/gds/puppet/blob/master/modules/mongodb/templates/configure-node-priority.js.erb#L26