Add node FQDN specific hiera data.
We need to be able to override some hieradata on
specific nodes.

Currently puppet has no knowledge of the ‘DR’
environment. Most applications are currently using
ENV variables that are set via puppet. These are
for things like `db_hostname`. We need to point
DR bouncer machines to read from the DR slave
database. As puppet doesn’t know about the DR env,
we need to do this on FQDN level hieradata.