Set screenshot as a service app absent
This app is being retired so set it to absent to remove it from servers.

Also remove the include phantomjs part - I did consider paramaterizing
the phantomjs class to allow passing ensure absent to the package
resource defined within, but that seems to be adding unneccessary
complexity, and it will be easier to remove the phantomjs package by
running a fabric task.