Backend vDC must be able to talk to Redis in API
sidekiq monitor resides in the backend vDC and has
to talk to Redis in the API vDC to get stats.

Preview needs a /8 as we haven’t renumbered.
Backend is 10.3.X.X and API layer is 10.7.X.X

Production and Staging have new numbering so we
can use a slightly less open subnet range of /16
as only the third octet changes between vDC.