Run worker for publishing-api
For performance reasons publishing-api puts everything that doesn't
have to happen in the request cycle on its Sidekiq queue.

In development we want to run this queue in 99% of cases, because we're
often depending on our request to be forwarded to the content-store.

The command to run the queue has been taken from publishing-api's own
Procfile.

The worker depends on the same things as publishing-api. When the V1
endpoints have been deleted we can probably remove publishing-api's
dependencies (except the worker).