Dry run network job: Don't delete .git directory
Configure the Jenkins GIT plugin not to delete the `.git` directory on
every build. The govuk-provisioning repo is approx. 100MB large because
it includes a binary artefact, so cloning the whole repository on each
Jenkins build is very time-consuming.

`wipe-workspace` is set to true by default. Instead, run `git-clean`
after `git-clone` to ensure that the workspace is clean.

See:
http://docs.openstack.org/infra/jenkins-job-builder/scm.html?highlight=force%20clone#scm.git

This was omitted in 1aa1725, which did the same for the actual job as
opposed to this 'dry-run' job.