Remove basic auth from Contacts Admin
We default to not enabling basic auth, but contacts admin was left protected
because of a concern that it might have unauthenticated routes for previewing
draft content, see: e73d8249815ad8069c52dd1b9dc5f34959503de2. This isn't the
case, so we can remove basic auth.

Contacts Admin is special in one way, and that is that the admin app is
deployed to frontend and backend, with configuration differences and is used to
serve the public index page:
https://www.gov.uk/government/organisations/hm-revenue-customs/contact