Fix Icinga service description for healthchecks
This service description doesn't mean the app is not running. The
app is normally running, it's just that the healthcheck is returning
an error.

Change the service description to reflect the service that is
actually being checked (the healthcheck). The status column
will contain information on the status of the service.