Specify owner and group of GPG mount
We provision Vagrant VMs in two ways:

- By running `govuk_puppet` from on the machine
- By running `vagrant provision` from the host machine

`govuk_puppet` has been broken because the puppetmaster process runs
as the puppet user, which did not have access to read the GPG mount
(it's owned by `vagrant:vagrant` by default).

Changing the owner and group to the puppet user fixes the issue because
`vagrant provision` runs as the vagrant user but uses sudo, so it can
read the directory contents anyway.

This issue was the one that prompted us to look at file permissions:
https://github.com/sihil/hiera-eyaml-gpg/issues/10