Refactor internal_tld hieradata
This is almost always set to app_domain. Once the preview environment
is deleted we can remove internal_tld entirely in favour of app_domain.

This commit changes the internal_tld in development and vagrant from
`development` to `dev.gov.uk`, but I don't anticipate any problems
as a result of that.