Remove `govuk_jenkins::config::users`
`users` is so similar to `admins` that there's no point in having both.

While tidying up, this commit removes the `Computer` permissions from
the half a dozen admins (which is no big deal because there are no
Jenkins slaves on the deploy Jenkins).

It also grants the `Item.Cancel`, `Item.Discover` and `View.Read` permissions
to the `ci_alphagov` user, which it didn't have before (the other admins did
because they're also in the 2ndline group). I don't consider this to be
a problem.