Only run jenkins-job-builder if Jenkins is up
jenkins-job-builder will try to exec at the end of every Puppet run,
even when Jenkins is in the process of reloading because its config
has changed.

This commit ensures that jenkins-job-builder will only run if Jenkins
is already available and will stop the Puppet run from erroring in the
case where Jenkins is unavailable.