Add host entry for search-3 in Integration
This host does not exist Integration, to save money, but the host entry
is needed to prevent Puppet from erroring when it evaluates the UFW
virtual resources in the `govuk::node::s_api_elasticsearch` class.

Add a FIXME as a reminder to fix the source of the problem at a later
date.