Add a passive check for configuring Bouncer's CDN service
This job has been failing without us noticing recently, due to authentication
errors on the Fastly account it uses. If we don't know it's failed then we may
tell departments that it's OK for them to transition domains before they've
been added to the CDN config, causing errors for users if the transition goes
ahead.

I've followed the recent PR to add a passive check for the
search-fetch-analytics-data job: https://github.gds/gds/puppet/pull/3645

This job should be triggered every night by the Transition_load_all_data job,
so setting the threshold to 104400 (29 hours) should catch an overnight
failure. The job also runs after Transition_load_site_config, which is
triggered by changes on master in the transition-config repo, and if it starts
erroring on one of those runs then we'll only know about it the next day (if I
understand passive checks correctly, which I'm not sure about).