Rename Production_{Redirector,Bouncer}_CDN job
We used to handle transitioned domains with a tool called Redirector, but
Bouncer replaced this a long time ago. Update the job name to refer to the
current application (the CDN service name already refers to Bouncer).

We'll lose the history of the job with this change, but there isn't much there
of use since the migration to Carrenza anyway.