[108068574] Add Launch VMs Licensify Job
Licensify lives in a different organistaion in the different
environments, so a separate job is the easiest to set this up with. We
have also never deployed Licensify/EFG in Skyscape (because they were
already running) so this gives us the option to do that in the future if
we wish to do so.