Add a warning about TLS settings
We're about to lock down our CDN configuration so that TLS connections
use only TLS v1.2 and our preferred cipher,
`ECDHE-RSA-AES256-GCM-SHA384`.

So that we don't change our supported protocols or ciphers and cause an
outage if they're not compatible with our Nginx configuration, add a
warning as a comment.