Always use SNI when checking SSL certs
`check_http` supports an `--sni` option which enables Server Name
Indication (SNI).

Enabling this option allows us to check distinct certificates that are
served from the same IP address.

Because the `server_name`, as supplied in the TLS hello during TLS
negotitation, can be different to the hostname or IP address that the
we connect to using TCP, we supply an additional argument to
`check_ssl_cert`.

I've updated all occurrences of `check_ssl_cert` accordingly. This fixes
a bug whereby we've been checking the expiry date of a snakeoil
certificate served on the default Nginx vhost of our mirrors, rather
than the certificate that is actually used to serve content.

I'm about to make use of this however to add a check to verify the two
certificates we serve at origin, `*.publishing.service.gov.uk` and
`www.gov.uk`, which does require SNI.

I've confirmed in the `check_http` source code that the value of the
`-I` option passed to `check_http` overrides the hostname set by `-H`,
which would otherwise determine which host to connect to over TCP:

https://github.com/nagios-plugins/nagios-plugins/blob/3de252dbee7066967269023c6d9215813a62dcdd/plugins/check_http.c#L518-L529