Explicitly pass through Nginx variables
When using Puppet 3 these parameters were automatically and magically passed through
from the parent class into the child class.

However when using the Puppet future parser (and therefore Puppet 4) the parameter
is not passed through which (for example) causes the Nginx config file to look like this:

```
    variables_hash_max_size ;
```

rather than this:

```
    variables_hash_max_size 512;
```

Which breaks all of our machines that run Nginx.

Explicitly passing through the parameters prevents this from happening.