Increase wal_keep_segments to 256
This value is the minimum number of 16MB segments that will be kept on
the PostgreSQL primary [1].

Since we've added standby machines on the other side of a VPN connection,
we keep seeing the data replication fall behind and stop replicating because
VPNs are unreliable.

Increasing the number of segments from 128 to 256 will mean that 4GB of disk
space (256 * 16MB) is used rather than 2GB.

We have plenty of space on both the root partition and the PostgreSQL data partition
on all of our PostgreSQL primary machines and I can't see any reason that we shouldn't
increase the number of segments we're storing.

I've chosen 256 because it's a nice round number which is double the current value.
If this doesn't fix the VPN replication issue we can look at increasing the number of
segments further or find an alternative fix.

[1]: http://www.postgresql.org/docs/9.3/static/runtime-config-replication.html#GUC-WAL-KEEP-SEGMENTS