Rework Jenkins Jobs namespacing
This commit removes environment specific Jenkins Job directories and
updates the relevant hieradata.

When I created a bunch of Jenkins Jobs originally I put them into
seperate directories to separate jobs that have specific variables for
different environments. This is not required because we already specify
jobs for different environments in the environment specific hieradata.