puppet-lint plugin for non-ASCII chars in comments
Add a `puppet-lint` plugin for non-ASCII characters found in comments in
Puppet manifests.

Example output:

    $ bundle exec rake lint
    ---> Running lint checks
    modules/govuk/manifests/apps/email_campaign_api.pp - WARNING: Non-ASCII character found in comment on line 21
    rake aborted!

    /Users/matt/govuk/puppet/lib/tasks/lint.rake:27:in `block in <top (required)>'
    Tasks: TOP => lint
    (See full trace by running task with --trace)

We only check Puppet comments, because Puppet's own parser validation
already catches non-ASCII characters in manifests outside of comments:

    $ bundle exec rake lint
    ---> Running lint checks
    manifests/site.pp - ERROR: Syntax error (try running `puppet parser validate <file>`) on line 2

It's important to catch these characters as they cause Puppet to fail
during the Puppet run:

    'Could not retrieve catalog from remote server: Error 400 on SERVER: invalid byte sequence in US-ASCII at
    /usr/share/puppet/production/current/modules/govuk/manifests/node/s_transition_postgresql_master.pp:1
    on node transition-postgresql-master-1.backend.production'

The `invalid byte sequence in US-ASCII` message is a Ruby error.

It seems that this mistake is relatively common because we all use Macs,
and pressing `alt` and `space` simultaneously in Mac OS X will create a
non-breaking space, which triggers the Ruby error.

This is inspired by:
- https://github.gds/gds/puppet/pull/2255
- 63b36f93bf75a848e2125008aa1e880c5861cf46
- 9c5943f1d9e4f2c69929db67e631f833c6730d2f
- 6ce832168a31d8794eafa23d09c7f2d8bc59e46b