Make vagrant hieradata more real
The monitoring user needs to be named 'nagios' because that name is
hardcoded in the defined type which grants the user permission.

The snakeoil key and certificate needs to be a real certificate and key
because otherwise PostgreSQL will fail to start. Note that this certificate
and key is not used for anything sensitive and is only used in the vagrant
environment. It's one I've just created.