Update common host hieradata
In order to start using vagrant definitions from the integration environment
rather than the preview environment we need to ensure that the host entries
match the networking config of the virtual machines.

This commit updates the common host hieradata so that it can be shared
between vagrant and integration. The preview specific host config is
moved to the preview file.

This requires adding a host entry for `email-campaign-mongo-1` in `common.yaml`
even though that machine isn't common (it doesn't exist in integration). This
is required to let the catalog compile for the node specs.