Hard code alpha taxonomy file location in a Rummager envvar
The Rummager migrate index task needs to know where to look for the alpha
taxonomy import file. This change ensures it points to the location used
by the Jenkins job which downloads the file in the first place. This makes
the alpha taxonomy available for prototyping with in the short-term, and
will be cleaned up/replaced by a more long-term solution that's currently
work in progress.