Improve whitehall attachment processing
Improve the processing for uploaded attachments so that they are
moved to the asset-slave instantly.

This is part of our disaster recovery work.

This change has been done in such a way that we can add a second
asset slave machine in a different datacentre which will receive
the attachments at the same time as the first asset slave.

In a later commit there will be a nightly rsync job which will ensure
that all of the files have been copied over (eg to mitigate the case
where the processing fails to copy one attachment because the network
is down, or the Puppetmaster is down (so govuk_node_list fails, etc).