Use keepLastValue() in PostgreSQL repl. lag graph
Use the Graphite `keepLastValue()` function when calculating the
PostgreSQL replication lag:

> Takes one metric or a wildcard seriesList. Continues the line with the
> last received value when gaps (‘None’ values) appear in your data,
> rather than breaking your line.

http://graphite.readthedocs.org/en/1.0/functions.html#graphite.render.functions.keepLastValue

This has the advantage that if the standby server's metrics are
unavailable for a period, e.g. if the VPN which the metrics are sent
over is down, the graph shows the difference between the primary's
current position and the last known position of the standby, rather than
showing a bogus spike showing the difference between the primary's
current and a zero value in place of the standby's position.

Should the reverse happen, and the primary's position is unknown but the
standy's position is known, the `transformNull(removeBelowValue())`
expression will ensure that values less than zero will appear as zero,
i.e. the standby is not behind the primary.

The `transformNull(removeBelowValue())` is also necessary to ensure that
the value reported is never negative, for example due to different
sampling times on the primary versus the standby.

Examples of graphs using this change:

Before:
https://graphite.publishing.service.gov.uk/render/?width=600&height=300&colorList=red,orange,blue,green,purple,brown&target=alias(dashed(constantLine(16777216)),%22critical%22)&target=alias(dashed(constantLine(8388608)),%22warning%22)&target=movingMedian(transformNull(removeBelowValue(diffSeries(postgresql-primary-1_backend*.postgresql-global.bytes-xlog_position,postgresql-standby-2_backend.postgresql-global.bytes-xlog_position),0),0),5)&from=20151201&until=20151207

After:
https://graphite.publishing.service.gov.uk/render/?width=600&height=300&colorList=red,orange,blue,green,purple,brown&target=alias(dashed(constantLine(16777216)),%22critical%22)&target=alias(dashed(constantLine(8388608)),%22warning%22)&target=movingMedian(transformNull(removeBelowValue(diffSeries(keepLastValue(postgresql-primary-1_backend*.postgresql-global.bytes-xlog_position),keepLastValue(postgresql-standby-2_backend.postgresql-global.bytes-xlog_position)),0),0),5)&from=20151201&until=20151207

Example graph that includes the period before which the PostgreSQL WAL
XLOG position was known, on a newly provisioned PostgreSQL primary and
standby:
https://graphite.publishing.service.gov.uk/render/?width=600&height=300&colorList=red,orange,blue,green,purple,brown&target=alias(dashed(constantLine(16777216)),%22critical%22)&target=alias(dashed(constantLine(8388608)),%22warning%22)&target=movingMedian(transformNull(removeBelowValue(diffSeries(keepLastValue(postgresql-primary-1_backend*.postgresql-global.bytes-xlog_position),keepLastValue(postgresql-standby-2_backend.postgresql-global.bytes-xlog_position)),0),0),5)&from=20141207&until=20151207