Run full attachment virus scan every hour
The last 3 runs of this virus scan in production have taken:

- 39 hours
- 40 hours
- 36 hours

As this command already runs with setlock there's no disadvantage to
making it run more frequently.

In its current state, if the run took 49 hours (ie just over 2 days)
we would have to wait almost another day for the run to start again.