Remove assets user from most machines
The assets user is the owner of the `/mnt/uploads` directory on the
asset class of machines. This means that the user account is also required on
the backend machines which make use of an NFS mount to serve attachments.

However the user is created on all machines because it's included in
`govuk::node::s_base`.

This isn't necessary because the `assets` class is specifically to set up a
machine for serving attachments. The apps that need it include the assets
class. Currently this is:

- asset_manager
- whitehall

I've moved the `include assets` out of the conditional in the whitehall
class because I think it's required on both the whitehall frontends and
backends. Either way this won't have be a negative change because it's already
included on all those machines through `govuk::node::s_base`.

I've also removed the include from `s_asset_slave` because it's already in
`s_asset_base` and from `s_base` because it shouldn't be there.