Run the monitor_logs script directly
Previously, govuk-cdn-logs-monitor was using the "startup.sh" shell
script to run the process which monitors logs.  However, when restarting
this script, the child ruby process is often left behind, and this isn't
picked up by monitoring.  The script simply runs "bundle exec ruby ...",
so instead run that command directly.  This should ensure that it is
cleaned up appropriately.