Disable publishing of topics for HMRC manuals
We added this feature to the HMRC manuals API earlier this year and it's still
behind a feature flag. HMRC aren't able to publish manuals as minor updates,
and so each publishing of a manual triggers email alerts for subscribers to
the topics that that manual is tagged to. For most manuals they've published
so far the initial publishing has failed several times before all sections
are published so they have had to retry, which has resulted in multiple emails
being sent to subscribers in a single day. Their system publishes the manual
before the sections, and they're unable to change this ordering of requests.

We've agreed to disable the publishing of topics until they're able to publish
manuals as minor updates to prevent this happening again.