Move CDN logs processed dir to separate disk
We've created a new disk to use for processing the CDN logs data to
prevent the boot partition or the data partition from filling up while
the processing is in progress.

This is at /mnt/logs_cdn_processed, but setting appropriate permissions
on that is fiddly, so instead make a subdirectory called `data` to hold
the data.