Configure disaster recovery MongoDB as hidden
> Hidden members are part of a replica set but cannot become primary
> and are invisible to client applications.

When we added priority 0 nodes across the disaster recovery VPN we discovered
that by default some MongoDB drivers will autodiscover all MongoDB replica
set members, including those that aren't specified in the application
configuration.

This led to application outages because applications would attempt to connect
to a replica set member which was unavailable because the VPN was down.

The second part of [the above quote][1] is important: we want the disaster recovery
members to not be used by the applications while the site is behaving normally.

This commit introduces a bit of duplication in the JavaScript but I'm not sure
it's worth refactoring now.

[1]: https://docs.mongodb.org/v2.4/core/replica-set-hidden-member/