Configure assets to serve from multiple hostnames
We're going to move the assets hostname to be under a `service.gov.uk`
subdomain, so we need to be able to configure both the old assets hostname
and the new assets hostname to respond at the same time.

We don't need the spec hieradata anymore because there's now a default
on the class.