Feature-flag service-manual-publisher publishing actions
We intend to deploy the application to production so that content designers can
collaborate on draft content. They can not do this on preview because of the
nightly data reset. We need to accumulate enough content in draft stage before
we can start publishing it, so we want to disable the publishing action [1] to
avoid accidental publishing.

This commit is expected to be reverted as soon as we're ready to publish.

[1] - https://github.com/alphagov/service-manual-publisher/pull/104